Update deflate.cpp
My first pull request ever on here, so apologies in advance if I'm not doing this right. Just corrected what I'm pretty sure was a very simple typo - removing the extra ", 1" on the end there makes the package compile again on all systems, so unless I'm wrong (and I've never programmed in any version of C/C++, so I could be - I'm only troubleshooting the error that gcc led me directly to), I think its inclusion there on the end was likely a mistake. Thanks for all the time you spend updating ECT - it's a fantastic tool, and I make sure to mention it when and wherever I get the chance.