Fix build compilation on Linux systems.
Linux systems need to have the -pthread flag explicitly stated in the compiler flags for threading. For mozjpeg, the sources are configured to use specific versions of e.g. aclocal which does not match the version installed in all systems, the source needs to be reconfigured for different systems prior to compiling.

While it should make no difference on the OSX side, please ensure that compiling keeps working as expected on OSX before merging, thank you.