Make file renames atomic
Unlinking the source file before renaming can lead to a situation
where, due to some error or ECT being interrupted by the user,
the unlink has executed but the rename has not, so no file at all
exists with the original file name. The recompressed file probably
exists at this point, but if the user is not careful (especially
if they are calling ECT from a script), they may delete it and end
up losing the file. Even if they do not delete it, various errors
(in case of scripting) and misunderstandings can occur because
the user expects their file to be available by its original name.

The only reason to unlink a file before renaming over it is that
on Windows, the `rename` function does not allow the destination
name to exist. However, Windows NT provides another API that can
be used to atomically rename and replace the destination file
if it exists. On other operating systems, `rename` already has
this behavior.

Instead of unlinking before renaming, use the appropriate API on
each platform that renames and replaces the destination in one go.

Caveat: Windows 95, 98 and ME do not have `MoveFileEx`.
After this commit, ECT no longer supports those operating systems.