feat: Support using "--" as an end-of-options marker
Following Unix argument passing conventions, specifying "--"
by itself as a command line argument signifies that every argument that
follows should be treated as positional rather than as an option,
even if it starts with a dash. This allows for easily passing filenames
as arguments that have names beginning with dashes.